# contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "School Management API",
...
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/roles": {
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "Create a role",
        "description": "Spec 6.1.4: name, description and at least one privilege code. Rejects the reserved name `Super Admin`, case-insensitive, and any privilege code that does not resolve (after legacy `guardian.*` alias resolution) to the register, naming the offender. Spec 6.1.7 rule 2: every requested privilege must already be held by the caller — a new role starts with none, so every one requested counts as an addition. `Idempotency-Key` is REQUIRED: a retry with the same key returns the same role instead of creating a second one.",
        "operationId": "CreateRole",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "The value of the __Host-XSRF-TOKEN cookie, echoed verbatim (double-submit CSRF, approved contract delta §5). Obtain it from GET /auth/csrf or from a prior response's Set-Cookie.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated key (UUID v4 recommended), 1-255 visible ASCII characters, no whitespace. Required on this route.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "List roles",
        "description": "Cursor-paginated per spec 9.5 — never offset. Archived roles are excluded unless `status` names them explicitly (spec 9.4's default-scope rule). `search` matches the role name, case-insensitively, by substring. `sort` is `name` (default) or `status`; `direction` is `asc` (default) or `desc`. `pageSize` defaults to 25 and is capped at 100.",
        "operationId": "ListRoles",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/RoleStatus"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageOfRoleDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRoleCommand": {
        "required": [
          "name",
          "description",
          "privileges"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "1..60 characters.",
            "example": "Class Teacher"
          },
          "description": {
            "type": [
              "null",
              "string"
            ],
            "description": "0..300 characters, or `null`.",
            "example": "Enters marks and views pupil records for an assigned arm."
          },
          "privileges": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "At least one privilege code. Legacy `guardian.*` aliases are accepted and resolved to their\ncanonical replacement before storage.",
            "example": [
              "result.score.enter",
              "pupil.view"
            ]
          }
        },
        "additionalProperties": false,
        "description": "`POST /api/v1/roles` (spec 6.1.4; approved delta\n        `.agent/decisions/2026-Q3-contract-deltas.md` entry `TASK-0028` §2). Rejects the reserved\n        name `Super Admin`, case-insensitive, and an unknown privilege code naming the offender.\n        Spec 6.1.7 rule 2 applies to every requested privilege (the role starts with none, so every one\n        requested is an \"add\") — enforced by the handler, not this type.",
        "example": {
          "name": "Class Teacher",
          "description": "Enters marks and views pupil records for an assigned arm.",
          "privileges": [
            "result.score.enter",
            "pupil.view"
          ]
        }
      },
...
      "CursorPageOfRoleDto": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            },
            "description": "The page of items, newest first. Empty (never null) when there is nothing more to return.",
            "example": [
              {
                "id": "0192f0c4-7c3e-7a1b-9f2d-3b8e5a6c1d40",
                "name": "Class Teacher",
                "description": "Enters marks and views pupil records for an assigned arm.",
                "isSystem": false,
                "privileges": [
                  "pupil.view",
                  "result.score.enter"
                ],
                "status": "Active"
              }
            ]
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ],
            "description": "`null` when this is the last page.",
            "example": "Y2xhc3MgdGVhY2hlch8wMTkyZjBjNC03YzNlLTdhMWItOWYyZC0zYjhlNWE2YzFkNDA="
          }
        },
        "additionalProperties": false,
        "description": "The cursor-pagination response envelope (spec 9.5). string? CursorPage&lt;TItem&gt;.NextCursor is opaque to the\nclient — it must be echoed back verbatim as the next request's cursor, and never parsed or\nconstructed by hand.",
        "example": {
          "items": [
            {
              "id": "0192f0c4-7c3e-7a1b-9f2d-3b8e5a6c1d40",
              "name": "Class Teacher",
              "description": "Enters marks and views pupil records for an assigned arm.",
              "isSystem": false,
              "privileges": [
                "pupil.view",
                "result.score.enter"
              ],
              "status": "Active"
            }
          ],
          "nextCursor": "Y2xhc3MgdGVhY2hlch8wMTkyZjBjNC03YzNlLTdhMWItOWYyZC0zYjhlNWE2YzFkNDA="
        }
      },
...
      "HttpValidationProblemDetails": {
        "required": [
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ],
            "example": "urn:schoolmanagement:error:sample_record.label_taken"
          },
          "title": {
            "type": [
              "null",
              "string"
            ],
            "example": "Conflict with current state"
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32",
            "example": 409
          },
          "detail": {
            "type": [
              "null",
              "string"
            ],
            "example": "A record with that label already exists."
          },
          "instance": {
            "type": [
              "null",
              "string"
            ],
            "example": "/api/v1/reference/records"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "Label": [
                "Label is required."
              ],
              "PageSize": [
                "PageSize must be at most 100."
              ]
            }
          },
          "errorCode": {
            "type": "string",
            "description": "Stable, machine-readable error code. Clients branch on this, never on `detail`. Absent when this response was produced directly by the framework rather than by this API's own result mapping."
          },
          "traceId": {
            "type": "string",
            "description": "Correlation id for this specific response occurrence. Present on every error response; quote it when reporting a problem."
          }
        },
        "additionalProperties": false,
        "description": "An RFC 9457 problem response for a validation failure, returned with status 422. Extends the standard problem shape with `errors`: an object keyed by request property name, whose values are the messages for that property, suitable for attaching to form fields. A 400 (rather than 422) means the request itself could not be parsed.",
        "example": {
          "type": "urn:schoolmanagement:error:request.validation_failed",
          "title": "Validation failed",
          "status": 422,
          "detail": "One or more validation errors occurred.",
          "instance": "/api/v1/reference/records",
          "errorCode": "request.validation_failed",
          "traceId": "0af7651916cd43dd8448eb211c80319c",
          "errors": {
            "Label": [
              "Label is required."
            ],
            "PageSize": [
              "PageSize must be at most 100."
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "required": [
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ],
            "example": "urn:schoolmanagement:error:sample_record.label_taken"
          },
          "title": {
            "type": [
              "null",
              "string"
            ],
            "example": "Conflict with current state"
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32",
            "example": 409
          },
          "detail": {
            "type": [
              "null",
              "string"
            ],
            "example": "A record with that label already exists."
          },
          "instance": {
            "type": [
              "null",
              "string"
            ],
            "example": "/api/v1/reference/records"
          },
          "errorCode": {
            "type": "string",
            "description": "Stable, machine-readable error code. Clients branch on this, never on `detail`. Absent when this response was produced directly by the framework (for example a model-binding failure or an authentication challenge) rather than by this API's own result mapping."
          },
          "traceId": {
            "type": "string",
            "description": "Correlation id for this specific response occurrence. Present on every error response; quote it when reporting a problem."
          },
          "lockedUntil": {
            "type": "string",
            "description": "UTC time the account's lockout ends (spec 6.1.11: five failed attempts locks it for fifteen minutes). Present only on the `423 Locked` response `POST /auth/sign-in` returns when the SUBMITTED password is correct but the account is currently locked — never on any other problem response.",
            "format": "date-time",
            "example": "2026-08-03T09:30:00+00:00"
          }
        },
        "additionalProperties": false,
        "description": "An RFC 9457 problem response. Returned for every error. Branch on the `errorCode` extension member — it is stable — and never on `detail`, which is human-readable prose that may be reworded. `traceId` identifies this specific occurrence in the server logs; quote it when reporting a problem. `type` is a stable URN of the form `urn:schoolmanagement:error:<code>`.",
        "example": {
          "type": "urn:schoolmanagement:error:sample_record.label_taken",
          "title": "Conflict with current state",
          "status": 409,
          "detail": "A record with that label already exists.",
          "instance": "/api/v1/reference/records",
          "errorCode": "sample_record.label_taken",
          "traceId": "0af7651916cd43dd8448eb211c80319c"
        }
      },
...
      "RoleDto": {
        "required": [
          "id",
          "name",
          "description",
          "isSystem",
          "privileges",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Opaque identifier.",
            "example": "0192f0c4-7c3e-7a1b-9f2d-3b8e5a6c1d40"
          },
          "name": {
            "type": "string",
            "description": "1..60 characters.",
            "example": "Class Teacher"
          },
          "description": {
            "type": [
              "null",
              "string"
            ],
            "description": "0..300 characters, or `null`.",
            "example": "Enters marks and views pupil records for an assigned arm."
          },
          "isSystem": {
            "type": "boolean",
            "description": "True only for the seeded Super Admin role. A system role cannot be edited, renamed, deleted or\nhave privileges removed (spec 6.1.4).",
            "example": false
          },
          "privileges": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Canonical codes, at least one, sorted deterministically.",
            "example": [
              "pupil.view",
              "result.score.enter"
            ]
          },
          "status": {
            "description": "RoleStatus.Archived roles are excluded from the default list (spec 9.4) but remain\n        individually readable and editable (except by `PATCH`/`DELETE` only when\n        IsSystem is true).",
            "$ref": "#/components/schemas/RoleStatus"
          }
        },
        "additionalProperties": false,
        "description": "The wire shape of a role (spec 6.1.4; approved delta `.agent/decisions/2026-Q3-contract-deltas.md`\nentry `TASK-0028` §2). Returned by every role endpoint — create, get, list (as the item\nshape) and update all share this one DTO, matching how `AdminAccountDetailDto` is reused\nacross its own family.",
        "example": {
          "id": "0192f0c4-7c3e-7a1b-9f2d-3b8e5a6c1d40",
          "name": "Class Teacher",
          "description": "Enters marks and views pupil records for an assigned arm.",
          "isSystem": false,
          "privileges": [
            "pupil.view",
            "result.score.enter"
          ],
          "status": "Active"
        }
      },
      "RoleStatus": {
        "enum": [
          "Active",
          "Archived"
        ],
        "description": "Lifecycle state of a Role (spec 6.1.4). \"An archived role cannot be newly assigned\nbut existing assignments continue until the session ends.\"",
        "example": "Active"
      },
...
    },
    "securitySchemes": {
      "CookieSession": {
        "type": "apiKey",
        "description": "Opaque session token set by POST /api/v1/auth/sign-in. HttpOnly — the browser attaches it automatically; it is never readable or settable from JavaScript.",
        "name": "__Host-Session",
        "in": "cookie"
      }
    }
  },
...
}